Rename file to use kebab-case
