+++ src/components/page-layout/components/example-index.js
